Refactor widgets into separate files
