+++ dask_ctl/tui/__init__.py
@@ -1,4 +1,5 @@
 try:
     from .tui import DaskCtlTUI
-except (ImportError, ModuleNotFoundError):
+except (ImportError, ModuleNotFoundError) as e:
+    raise e
     DaskCtlTUI = None
