+++ dask_ctl/tui/widgets/info.py
@@ -1,0 +1,28 @@
+import sys
+
+from rich.text import Text
+from textual.widget import Widget
+import dask
+import distributed
+
+
+class Info(Widget):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+        from ... import __version__
+
+        self.versions = {
+            "python": sys.version.split("|")[0].strip(),
+            "dask": dask.__version__,
+            "distributed": distributed.__version__,
+            "dask-ctl": __version__,
+        }
+
+    def render(self) -> Text:
+        outs = []
+        outs.append(("Dask Control\n", "bold blue"))
+        for version in self.versions:
+            outs.append((f"{version}: ", "bold orange3"))
+            outs.append(f"{self.versions[version]}\n")
+        return Text.assemble(*outs)
